Add Ui, Parser and Sotrage classes
