+++ data/hewwoTasks.txt
@@ -1,3 +1,4 @@
 *************************************************
-1. [D][ ] die (by: Oct 15 2019)
+1. [D][ ] (by: Oct 15 2020)
+2. [E][ ] cry (from: 4pm to: 6pm)
 *************************************************
